requirements.txt
torch>=2.5.0
transformers>=4.40.0
datasets>=3.0.0
faiss-cpu
vllm>=0.6.0
numpy>=1.24.0
tqdm
